+++ packages/react/src/components/tooltip/index.ts
@@ -1,0 +1,7 @@
+export {
+  Tooltip,
+  TooltipContent,
+  TooltipPortal,
+  TooltipProvider,
+  TooltipTrigger,
+} from "./tooltip";